Create table for user, tourney & match

+++ db/migrate/20211021151440_create_matches.rb
@@ -2,8 +2,8 @@ class CreateMatches < ActiveRecord::Migration[6.1]
   def change
     create_table :matches do |t|
       t.string :mp_link
-      t.int :warmup
-      t.double :matchcost
+      t.integer :warmup
+      t.decimal :matchcost
       t.bigint :average_score
 
       t.timestamps
